+++ yarn-project/end-to-end/src/single-node/cross-chain/l1_to_l2_inbox_drift.test.ts
@@ -1,38 +1,42 @@
-import { AztecAddress } from '@aztec/aztec.js/addresses';
+import type { AztecAddress } from '@aztec/aztec.js/addresses';
 import { generateClaimSecret } from '@aztec/aztec.js/ethereum';
 import { Fr } from '@aztec/aztec.js/fields';
 import type { Logger } from '@aztec/aztec.js/log';
-import { isL1ToL2MessageReady } from '@aztec/aztec.js/messaging';
 import type { AztecNode } from '@aztec/aztec.js/node';
 import { TxExecutionResult } from '@aztec/aztec.js/tx';
 import type { Wallet } from '@aztec/aztec.js/wallet';
-import { BlockNumber } from '@aztec/foundation/branded-types';
+import type { BlockNumber } from '@aztec/foundation/branded-types';
 import { timesAsync } from '@aztec/foundation/collection';
 import { retryUntil } from '@aztec/foundation/retry';
 import { TestContract } from '@aztec/noir-test-contracts.js/Test';
-import { ExecutionPayload } from '@aztec/stdlib/tx';
 
 import { jest } from '@jest/globals';
 
 import { L1_DIRECT_WRITE_ACCOUNT_INDEX, PIPELINING_SETUP_OPTS } from '../../fixtures/fixtures.js';
-import { sendL1ToL2Message } from '../../fixtures/l1_to_l2_messaging.js';
 import { waitForBlockNumber } from '../../fixtures/wait_helpers.js';
 import { CrossChainMessagingTest } from './cross_chain_messaging_test.js';
+import { type L1ToL2MessageScope, createL1ToL2MessageHelpers } from './message_test_helpers.js';
 
 jest.setTimeout(300_000);
 
-// L1→L2 messaging via Inbox: message readiness, duplicate messages, and inbox drift after a rollup
-// reorg. Uses CrossChainMessagingTest (prod sequencer, pipelining preset: ethSlot=4s, aztecSlot=12s,
-// inboxLag=2, minTxsPerBlock=1, aztecProofSubmissionEpochs=2, aztecEpochDuration=4) with
-// EpochTestSettler for auto-proving and CrossChainTestHarness for L1↔L2 token portal bridging.
-// Each it is run as an independent CI job (*.parallel.test.ts convention).
-describe('single-node/cross-chain/l1_to_l2', () => {
+// L1→L2 messaging via Inbox: inbox checkpoint drift after a rollup reorg. Uses CrossChainMessagingTest
+// (prod sequencer, pipelining preset: ethSlot=4s, aztecSlot=12s, inboxLag=2, minTxsPerBlock=1,
+// aztecProofSubmissionEpochs=2, aztecEpochDuration=4) with EpochTestSettler for auto-proving and
+// CrossChainTestHarness for L1↔L2 token portal bridging. The drift scenario runs over private and
+// public scope via it.each, all sharing one node stood up once in beforeAll.
+describe('single-node/cross-chain/l1_to_l2_inbox_drift', () => {
   let t: CrossChainMessagingTest;
+
   let log: Logger;
   let aztecNode: AztecNode;
   let wallet: Wallet;
   let user1Address: AztecAddress;
   let testContract: TestContract;
+
+  let sendMessageToL2: ReturnType<typeof createL1ToL2MessageHelpers>['sendMessageToL2'];
+  let advanceBlock: ReturnType<typeof createL1ToL2MessageHelpers>['advanceBlock'];
+  let waitForMessageFetched: ReturnType<typeof createL1ToL2MessageHelpers>['waitForMessageFetched'];
+  let waitForMessageReady: ReturnType<typeof createL1ToL2MessageHelpers>['waitForMessageReady'];
 
   // Whether explicit mark-as-proven calls are honored. The inbox-drift scenario flips this to
   // false to let the proposed chain drift and prune.
@@ -48,10 +52,9 @@ describe('single-node/cross-chain/l1_to_l2', () => {
     await t.cheatCodes.rollup.markAsProven();
   };
 
-  beforeEach(async () => {
-    markProvenEnabled = true;
+  beforeAll(async () => {
     t = new CrossChainMessagingTest(
-      'l1_to_l2',
+      'l1_to_l2_inbox_drift',
       // PIPELINING_SETUP_OPTS sets minTxsPerBlock=0; this test needs minTxsPerBlock=1 because it
       // manually mines blocks one tx at a time via advanceBlock() and counts checkpoints, so empty
       // pipelined checkpoints interleaving between txs would break the exact block-number assertions.
@@ -69,40 +72,30 @@ describe('single-node/cross-chain/l1_to_l2', () => {
 
     ({ logger: log, wallet, user1Address, aztecNode } = t);
     ({ contract: testContract } = await TestContract.deploy(wallet).send({ from: user1Address }));
+
+    ({ sendMessageToL2, advanceBlock, waitForMessageFetched, waitForMessageReady } = createL1ToL2MessageHelpers({
+      t,
+      aztecNode,
+      wallet,
+      user1Address,
+      log,
+      markAsProven,
+    }));
   }, 300_000);
 
-  afterEach(async () => {
+  // Reset the proving gate before every it so a scenario that disabled it can't leak into the next.
+  beforeEach(() => {
+    markProvenEnabled = true;
+  });
+
+  afterAll(async () => {
     await t.teardown();
   });
 
-  // Sends an L1→L2 message from the harness L1 account. This suite skips the token bridge, so the
-  // message context is built from the test's L1 handles rather than a CrossChainTestHarness.
-  const sendMessageToL2 = (message: { recipient: AztecAddress; content: Fr; secretHash: Fr }) =>
-    sendL1ToL2Message(message, {
-      l1Client: t.harnessL1Client,
-      l1ContractAddresses: t.deployL1ContractsValues.l1ContractAddresses,
-    });
-
-  const getConsumeMethod = (scope: 'private' | 'public') =>
+  const getConsumeMethod = (scope: L1ToL2MessageScope) =>
     scope === 'private'
       ? testContract.methods.consume_message_from_arbitrary_sender_private
       : testContract.methods.consume_message_from_arbitrary_sender_public;
-
-  // Sends a tx to L2 to advance the block number by 1
-  const advanceBlock = async () => {
-    const block = await aztecNode.getBlockNumber();
-    log.warn(`Sending noop tx at block ${block}`);
-    await wallet.sendTx(ExecutionPayload.empty(), { from: user1Address });
-    const newBlock = await aztecNode.getBlockNumber();
-    log.warn(`Advanced to block ${newBlock}`);
-    if (newBlock === block) {
-      throw new Error(`Failed to advance block ${block}`);
-    }
-    // Keep the proof window from expiring mid-test (see `markAsProven` above). No-op once the drift
-    // scenario disables proving.
-    await markAsProven();
-    return newBlock;
-  };
 
   const waitForBlockToCheckpoint = async (blockNumber: BlockNumber) => {
     await waitForBlockNumber(aztecNode, blockNumber, { tag: 'checkpointed', timeout: 60 });
@@ -125,7 +118,7 @@ describe('single-node/cross-chain/l1_to_l2', () => {
     log.warn(`At checkpoint ${checkpoint}`);
   };
 
-  // Same as above but ignores errors. Useful if we expect a prune.
+  // Same as advanceBlock but ignores errors. Useful if we expect a prune.
   const tryAdvanceBlock = async () => {
     try {
       await advanceBlock();
@@ -134,115 +127,11 @@ describe('single-node/cross-chain/l1_to_l2', () => {
     }
   };
 
-  // Waits until the message is fetched by the archiver of the node and returns the msg target checkpoint.
-  // Advances a block on each retry because an L1->L2 message is only indexed once further L2 blocks build.
-  const waitForMessageFetched = async (msgHash: Fr) => {
-    log.warn(`Waiting until the message is fetched by the node`);
-    return await retryUntil(
-      async () => {
-        const checkpoint = await aztecNode.getL1ToL2MessageCheckpoint(msgHash);
-        if (checkpoint !== undefined) {
-          return checkpoint;
-        }
-        await advanceBlock();
-        return undefined;
-      },
-      'get msg checkpoint',
-      60,
-    );
-  };
-
-  // Waits until the message is ready to be consumed on L2 as it's been added to the world state
-  const waitForMessageReady = async (
-    msgHash: Fr,
-    scope: 'private' | 'public',
-    onNotReady?: (blockNumber: BlockNumber) => Promise<void>,
-  ) => {
-    const msgCheckpoint = await waitForMessageFetched(msgHash);
-    log.warn(
-      `Waiting until L2 reaches the first block of msg checkpoint ${msgCheckpoint} (current is ${await aztecNode.getCheckpointNumber()})`,
-    );
-    await retryUntil(
-      async () => {
-        const [blockNumber, checkpointNumber] = await Promise.all([
-          aztecNode.getBlockNumber(),
-          aztecNode.getCheckpointNumber(),
-        ]);
-        const witness = await aztecNode.getL1ToL2MessageMembershipWitness('latest', msgHash);
-        const isReady = await isL1ToL2MessageReady(aztecNode, msgHash);
-        log.info(
-          `Block is ${blockNumber}, checkpoint is ${checkpointNumber}. Message checkpoint is ${msgCheckpoint}. Witness ${!!witness}. Ready ${isReady}.`,
-        );
-        if (!isReady) {
-          await (onNotReady ? onNotReady(blockNumber) : advanceBlock());
-        }
-        return isReady;
-      },
-      `wait for rollup to reach msg checkpoint ${msgCheckpoint}`,
-      240,
-    );
-  };
-
-  // We register one portal address when deploying contract but that address is no-longer the only address
-  // allowed to send messages to the given contract. In the following test we'll test that it's really the case.
-  // We'll also test that we can send the same message content across the bridge multiple times.
-  const canSendMessageFromNonRegisteredPortal = async (scope: 'private' | 'public') => {
-    // Generate and send the message to the L1 contract
-    const [secret, secretHash] = await generateClaimSecret();
-    const message = { recipient: testContract.address, content: Fr.random(), secretHash };
-    const { msgHash: message1Hash, globalLeafIndex: actualMessage1Index } = await sendMessageToL2(message);
-
-    await waitForMessageReady(message1Hash, scope);
-
-    const [message1Index] = (await aztecNode.getL1ToL2MessageMembershipWitness('latest', message1Hash))!;
-    expect(actualMessage1Index.toBigInt()).toBe(message1Index);
-
-    const sendConsumeMsgTx = async (index: Fr) => {
-      const call = getConsumeMethod(scope)(message.content, secret, t.ethAccount, index);
-      if (scope === 'public') {
-        await call.simulate({ from: user1Address });
-      }
-      await call.send({ from: user1Address });
-    };
-
-    // We consume the L1 to L2 message using the test contract either from private or public
-    await sendConsumeMsgTx(actualMessage1Index);
-
-    // We send and consume the exact same message the second time to test that oracles correctly return the new
-    // non-nullified message
-    const { msgHash: message2Hash, globalLeafIndex: actualMessage2Index } = await sendMessageToL2(message);
-
-    // We check that the duplicate message was correctly inserted by checking that its message index is defined
-    await waitForMessageReady(message2Hash, scope);
-
-    const [message2Index] = (await aztecNode.getL1ToL2MessageMembershipWitness('latest', message2Hash))!;
-    expect(message2Index).toBeDefined();
-    expect(message2Index).toBeGreaterThan(actualMessage1Index.toBigInt());
-    expect(actualMessage2Index.toBigInt()).toBe(message2Index);
-
-    // Now we consume the message again. Everything should pass because oracle should return the duplicate message
-    // which is not nullified
-    await sendConsumeMsgTx(actualMessage2Index);
-  };
-
-  // Sends the same L1→L2 message content twice via a non-registered portal, waits for each to be
-  // ready, and consumes both from private. Verifies duplicate messages are indexed correctly and
-  // the second consumption uses the non-nullified duplicate leaf.
-  it('can send an L1 to L2 message from a non-registered portal address consumed from private repeatedly', async () => {
-    await canSendMessageFromNonRegisteredPortal('private');
-  });
-
-  // Same as above but the message is consumed from public state. Verifies the public consumption
-  // path handles duplicate messages and the oracle returns the correct non-nullified leaf index.
-  it('can send an L1 to L2 message from a non-registered portal address consumed from public repeatedly', async () => {
-    await canSendMessageFromNonRegisteredPortal('public');
-  });
-
   // Inbox checkpoint number can drift on two scenarios: if the rollup reorgs and rolls back its own
   // checkpoint number, or if the inbox receives too many messages and they are inserted faster than
   // they are consumed. In this test, we mine several checkpoints without marking them as proven until
   // we can trigger a reorg, and then wait until the message can be processed to consume it.
-  const canConsumeMessageAfterInboxDrift = async (scope: 'private' | 'public') => {
+  const canConsumeMessageAfterInboxDrift = async (scope: L1ToL2MessageScope) => {
     // Stop the background epoch test settler so the drift scenario below can proceed without
     // an auto-prover racing it.
     await t.epochTestSettler?.stop();
@@ -333,15 +222,13 @@ describe('single-node/cross-chain/l1_to_l2', () => {
   };
 
   // Mines four checkpoints without proving, inserting an L1→L2 message after the drift, then
-  // triggers a rollup prune back to the pre-drift block. Verifies the message can be consumed from
-  // private only after the chain re-syncs to the message's checkpoint, not before.
-  it('can consume L1 to L2 message in private after inbox drifts away from the rollup', async () => {
-    await canConsumeMessageAfterInboxDrift('private');
-  });
-
-  // Same drift scenario but consuming from public. Uses a send+dontThrowOnRevert loop to probe when
-  // the message becomes consumable, then verifies the successful tx is in the message's checkpoint.
-  it('can consume L1 to L2 message in public after inbox drifts away from the rollup', async () => {
-    await canConsumeMessageAfterInboxDrift('public');
-  });
+  // triggers a rollup prune back to the pre-drift block. Verifies the message can be consumed only
+  // after the chain re-syncs to the message's checkpoint, not before, from both private and public
+  // scope (public uses a send+dontThrowOnRevert loop to probe when the message becomes consumable).
+  it.each(['private', 'public'] as const)(
+    'can consume L1 to L2 message in %s after inbox drifts away from the rollup',
+    async scope => {
+      await canConsumeMessageAfterInboxDrift(scope);
+    },
+  );
 });
