--- yarn-project/end-to-end/src/single-node/cross-chain/token_bridge_public.parallel.test.ts
@@ -1,160 +1,0 @@
-import { Fr } from '@aztec/aztec.js/fields';
-
-import { jest } from '@jest/globals';
-
-import {
-  L1_DIRECT_WRITE_ACCOUNT_INDEX,
-  NO_L1_TO_L2_MSG_ERROR,
-  PIPELINING_SETUP_OPTS,
-} from '../../fixtures/fixtures.js';
-import { waitForL2ToL1Witness } from '../../fixtures/wait_helpers.js';
-import { CrossChainMessagingTest } from './cross_chain_messaging_test.js';
-
-// Public L1→L2 token deposit and L2→L1 withdrawal via the TokenBridge. Uses CrossChainMessagingTest
-// with startProverNode=true (prod sequencer, pipelining preset: ethSlot=4s, aztecSlot=12s), fake
-// in-proc prover node, and CrossChainTestHarness for full L1↔L2 portal/bridge lifecycle. Setup and
-// teardown happen per-test (beforeEach/afterEach) because the test creates fresh bridge state each run.
-describe('single-node/cross-chain/token_bridge_public', () => {
-  // Pipelining slows wall-clock chain progress (12s slots); waitForProven via advanceToEpochProven
-  // needs more than the default 300s per-test budget.
-  jest.setTimeout(15 * 60 * 1000);
-
-  const t = new CrossChainMessagingTest(
-    'token_bridge_public',
-    { startProverNode: true },
-    {},
-    {},
-    { l1HarnessAccountIndex: L1_DIRECT_WRITE_ACCOUNT_INDEX },
-  );
-
-  let { crossChainTestHarness, ethAccount, aztecNode, logger, ownerAddress, l2Bridge, l2Token, wallet, user2Address } =
-    t;
-
-  beforeEach(async () => {
-    await t.setup({ ...PIPELINING_SETUP_OPTS });
-    // Have to destructure again to ensure we have latest refs.
-    ({ crossChainTestHarness, wallet, user2Address } = t);
-
-    ethAccount = crossChainTestHarness.ethAccount;
-    aztecNode = crossChainTestHarness.aztecNode;
-    logger = crossChainTestHarness.logger;
-    ownerAddress = crossChainTestHarness.ownerAddress;
-    l2Bridge = crossChainTestHarness.l2Bridge;
-    l2Token = crossChainTestHarness.l2Token;
-  }, 300_000);
-
-  afterEach(async () => {
-    await t.teardown();
-  });
-
-  // Full round-trip: mint on L1, publicly deposit via TokenPortal, wait for message, claim_public on
-  // L2, authorize bridge to burn, withdraw to L1, advance to epoch proven, consume Outbox on L1.
-  // Asserts L1 balance is restored after the round-trip.
-  it('Publicly deposit funds from L1 -> L2 and withdraw back to L1', async () => {
-    const l1TokenBalance = 1000000n;
-    const bridgeAmount = 100n;
-
-    // 1. Mint tokens on L1
-    logger.verbose(`1. Mint tokens on L1`);
-    await crossChainTestHarness.mintTokensOnL1(l1TokenBalance);
-
-    // 2. Deposit tokens to the TokenPortal
-    logger.verbose(`2. Deposit tokens to the TokenPortal`);
-    const claim = await crossChainTestHarness.sendTokensToPortalPublic(bridgeAmount);
-    const msgHash = Fr.fromHexString(claim.messageHash);
-    expect(await crossChainTestHarness.getL1BalanceOf(ethAccount)).toBe(l1TokenBalance - bridgeAmount);
-
-    // Wait for the message to be available for consumption
-    logger.verbose(`Wait for the message to be available for consumption`);
-    await crossChainTestHarness.makeMessageConsumable(msgHash);
-
-    // Check message leaf index matches
-    const maybeIndexAndPath = await aztecNode.getL1ToL2MessageMembershipWitness('latest', msgHash);
-    expect(maybeIndexAndPath).toBeDefined();
-    const messageLeafIndex = maybeIndexAndPath![0];
-    expect(messageLeafIndex).toEqual(claim.messageLeafIndex);
-
-    // 3. Consume L1 -> L2 message and mint public tokens on L2
-    logger.verbose('3. Consume L1 -> L2 message and mint public tokens on L2');
-    await crossChainTestHarness.consumeMessageOnAztecAndMintPublicly(claim);
-    await crossChainTestHarness.expectPublicBalanceOnL2(ownerAddress, bridgeAmount);
-    const afterBalance = bridgeAmount;
-
-    // Time to withdraw the funds again!
-    logger.info('Withdrawing funds from L2');
-
-    // 4. Give approval to bridge to burn owner's funds:
-    const withdrawAmount = 9n;
-    const authwitNonce = Fr.random();
-    const validateActionInteraction = await wallet.setPublicAuthWit(
-      ownerAddress,
-      {
-        caller: l2Bridge.address,
-        action: l2Token.methods.burn_public(ownerAddress, withdrawAmount, authwitNonce),
-      },
-      true,
-    );
-    await validateActionInteraction.send();
-
-    // 5. Withdraw owner's funds from L2 to L1
-    logger.verbose('5. Withdraw owner funds from L2 to L1');
-    const l2ToL1Message = await crossChainTestHarness.getL2ToL1MessageLeaf(withdrawAmount);
-    const l2TxReceipt = await crossChainTestHarness.withdrawPublicFromAztecToL1(withdrawAmount, authwitNonce);
-    await crossChainTestHarness.expectPublicBalanceOnL2(ownerAddress, afterBalance - withdrawAmount);
-
-    // Advance the epoch until the tx is proven since the messages are inserted to the outbox when the epoch is proven.
-    await t.advanceToEpochProven(l2TxReceipt);
-
-    const l2ToL1MessageResult = await waitForL2ToL1Witness(aztecNode, l2TxReceipt.txHash, l2ToL1Message, {
-      timeout: 60,
-    });
-
-    // Check balance before and after exit.
-    expect(await crossChainTestHarness.getL1BalanceOf(ethAccount)).toBe(l1TokenBalance - bridgeAmount);
-    await crossChainTestHarness.withdrawFundsFromBridgeOnL1(
-      withdrawAmount,
-      l2ToL1MessageResult.epochNumber,
-      l2ToL1MessageResult.numCheckpointsInEpoch,
-      l2ToL1MessageResult.leafIndex,
-      l2ToL1MessageResult.siblingPath,
-    );
-    expect(await crossChainTestHarness.getL1BalanceOf(ethAccount)).toBe(l1TokenBalance - bridgeAmount + withdrawAmount);
-  }, 900_000);
-
-  // User2 tries to claim to their own address (fails), then correctly claims to ownerAddress.
-  // Asserts only ownerAddress receives the tokens, user2 gets nothing, and the message is consumed.
-  it('Someone else can mint funds to me on my behalf (publicly)', async () => {
-    const l1TokenBalance = 1000000n;
-    const bridgeAmount = 100n;
-
-    await crossChainTestHarness.mintTokensOnL1(l1TokenBalance);
-    const claim = await crossChainTestHarness.sendTokensToPortalPublic(bridgeAmount);
-    const msgHash = Fr.fromHexString(claim.messageHash);
-    expect(await crossChainTestHarness.getL1BalanceOf(ethAccount)).toBe(l1TokenBalance - bridgeAmount);
-
-    await crossChainTestHarness.makeMessageConsumable(msgHash);
-
-    // Check message leaf index matches
-    const maybeIndexAndPath = await aztecNode.getL1ToL2MessageMembershipWitness('latest', msgHash);
-    expect(maybeIndexAndPath).toBeDefined();
-    const messageLeafIndex = maybeIndexAndPath![0];
-    expect(messageLeafIndex).toEqual(claim.messageLeafIndex);
-
-    // user2 tries to consume this message and minting to itself -> should fail since the message is intended to be consumed only by owner.
-    await expect(
-      l2Bridge.methods
-        .claim_public(user2Address, bridgeAmount, claim.claimSecret, messageLeafIndex)
-        .simulate({ from: user2Address }),
-    ).rejects.toThrow(NO_L1_TO_L2_MSG_ERROR);
-
-    // user2 consumes owner's L1-> L2 message on bridge contract and mints public tokens on L2
-    logger.info("user2 consumes owner's message on L2 Publicly");
-    await l2Bridge.methods
-      .claim_public(ownerAddress, bridgeAmount, claim.claimSecret, messageLeafIndex)
-      .send({ from: user2Address });
-
-    // ensure funds are gone to owner and not user2.
-    await crossChainTestHarness.expectPublicBalanceOnL2(ownerAddress, bridgeAmount);
-    await crossChainTestHarness.expectPublicBalanceOnL2(user2Address, 0n);
-  }, 90_000);
-});
